Add files via upload
This is an algorithm that identifies the presence of different prefixes in last names that are in a csv file.  Once it has identified the prefix presence, then it select the compound last name from the complete last name.

There is an error that I'm still trying to fix, with the "DE" prefix which is being applied to parts of names that are not compounded.